ml_trading_system/requirements.txt
# ML Trading System Requirements

# Data & Exchange
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
xgboost>=2.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
torch>=2.0.0  # For future LSTM/Transformer models

# Technical Analysis
ta>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Config & Data
pyyaml>=6.0

# API & Server (for live trading)
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0

# Optional: GPU support for XGBoost
# xgboost[gpu]>=2.0.0

# Optional: Advanced monitoring
# mlflow>=2.0.0

